+++ test/support/mocks.ex
@@ -1,3 +1,4 @@
 Mox.defmock(ElasticsearchMock, for: Elasticsearch.API)
 Mox.defmock(MockClusterHandler, for: TdCluster.ClusterHandler)
 Mox.defmock(TdDd.Loader.MockWorker, for: TdDd.Loader.Worker.Behaviour)
+Mox.defmock(TdDd.Search.EnricherImplMock, for: TdDd.Search.EnricherBehaviour)