+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-6219]: 
+  - Put data structure embeddings in elastic
+  - Add suggestion controller
+
+### Fixed
+
+[TD-7226] Enhance SSL configuration handling in production
+
 ## [7.5.0] 2025-04-30
 
 ### Added